# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText API
  description: API per la piattaforma di messaggistica WASAText - Homework 1
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Server di sviluppo locale

paths:
  /session:
    post:
      summary: Login o registrazione utente
      operationId: doLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: "Maria"
      responses:
        "201":
          description: Login riuscito
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    example: 12345
        "400":
          $ref: '#/components/responses/BadRequest'
  /me/username:
    put:
      tags:
        - User
      summary: Cambia il proprio username
      operationId: setMyUserName
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "MariaNuovo"
      responses:
        "204":
          description: Nome cambiato con successo
        "401":
          $ref: '#/components/responses/Unauthorized'
  /me/picture:
    put:
      tags:
        - User
      summary: Imposta la foto profilo
      operationId: setMyPhoto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [picture]
              properties:
                picture:
                  type: string
                  description: L'immagine codificata in Base64
                  example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAE..."
      responses:
        "204":
          description: Foto aggiornata con successo
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /users:
    get:
      tags:
        - User
      summary: Cerca utenti
      description: Serve per trovare persone da aggiungere a una nuova chat.
      operationId: getUsers
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          description: Stringa di ricerca per username
          schema:
            type: string
            example: "Mar"
      responses:
        "200":
          description: Risultati ricerca
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /conversations:
    get:
      tags:
        - Conversations
      summary: Ottieni la lista delle conversazioni
      description: Restituisce l'elenco di tutte le chat dell'utente in ordine cronologico inverso.
      operationId: getMyConversations
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista recuperata con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Conversations
      summary: Crea una nuova chat o gruppo
      description: Crea una nuova conversazione. Se specifichi un solo utente è una chat privata, altrimenti è un gruppo.
      operationId: createConversation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participants]
              properties:
                groupName:
                  type: string
                  description: Opzionale. Se presente, crea un gruppo con questo nome.
                  example: "Calcetto"
                participants:
                  type: array
                  description: Lista degli ID utenti da invitare
                  items:
                    type: integer
                  example: [55, 62]
      responses:
        "201":
          description: Conversazione creata con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId:
                    type: integer
                    example: 202
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /conversations/{conversationId}:
    get:
      tags:
        - Conversations
      summary: Leggi i messaggi di una conversazione
      description: Restituisce la cronologia completa dei messaggi per una specifica chat.
      operationId: getConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: L'ID della conversazione da visualizzare
          schema:
            type: integer
      responses:
        "200":
          description: Messaggi recuperati con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Conversations
      summary: Invia un messaggio
      description: Invia un nuovo messaggio (testo o foto) in una conversazione esistente.
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, type]
              properties:
                content:
                  type: string
                  description: Il testo del messaggio o l'identificatore dell'immagine.
                  example: "Ciao! Come va l'homework?"
                type:
                  type: string
                  enum: [text, image]
                  example: "text"
                replyToId:
                  type: integer
                  description: ID del messaggio a cui si sta rispondendo
                  example: 5001
      responses:
        "201":
          description: Messaggio inviato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/messages/{messageId}:
    delete:
      tags:
        - Conversations
      summary: Cancella un messaggio
      description: Rimuove un messaggio specifico da una conversazione.
      operationId: deleteMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Messaggio cancellato con successo
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/messages/{messageId}/comment:
    post:
      tags:
        - Conversations
      summary: Aggiungi una reazione a un messaggio
      operationId: commentMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: Il testo della reazione o l'emoji
                  example: "👍"
      responses:
        "201":
          description: Reazione aggiunta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Conversations
      summary: Rimuovi la reazione dal messaggio
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Reazione rimossa con successo
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/forward:
    post:
      tags:
        - Conversations
      summary: Inoltra un messaggio
      description: Inoltra un messaggio esistente verso questa conversazione.
      operationId: forwardMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: ID della conversazione in cui vuoi inoltrare il messaggio
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messageId]
              properties:
                messageId:
                  type: integer
                  description: L'ID del messaggio originale da inoltrare
                  example: 999
      responses:
        "201":
          description: Messaggio inoltrato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        "401":
          $ref: '#/components/responses/Unauthorized'  
  /conversations/{conversationId}/members/{userId}:
    put:
      tags:
        - Groups
      summary: Aggiungi utente al gruppo
      description: Aggiunge un partecipante a una conversazione di gruppo esistente.
      operationId: addToGroup
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          description: L'ID dell'utente da aggiungere
          schema:
            type: integer
      responses:
        "204":
          description: Utente aggiunto con successo
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Groups
      summary: Lascia il gruppo
      description: Rimuove l'utente specificato dai partecipanti del gruppo.
      operationId: leaveGroup
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Uscita dal gruppo avvenuta con successo
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/name:
    put:
      tags:
        - Groups
      summary: Cambia il nome del gruppo
      operationId: setGroupName
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Il nuovo nome del gruppo
                  example: "Calcetto Venerdì"
      responses:
        "204":
          description: Nome del gruppo aggiornato
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/photo:
    put:
      tags:
        - Groups
      summary: Imposta la foto del gruppo
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  description: L'immagine del gruppo codificata in Base64
                  example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
      responses:
        "204":
          description: Foto del gruppo aggiornata
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: La richiesta è malformata
    Unauthorized:
      description: Non autorizzato
    NotFound:
      description: Risorsa non trovata
    InternalServerError:
      description: Errore interno del server
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
          example: 55
        username:
          type: string
          example: "MarioRossi"
    Conversation:
      type: object
      properties:
        conversationId:
          type: integer
          example: 101
        name:
          type: string
          description: Nome dell'utente o del gruppo
          example: "Gruppo Studio"
        photo:
          type: string
          description: Immagine del profilo/gruppo in Base64
          example: "data:image/png;base64,..."
        lastMessage:
          type: string
          description: Anteprima dell'ultimo messaggio inviato
          example: "Ci vediamo alle 15?"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-04T15:30:00Z"
    Message:
      type: object
      properties:
        messageId:
          type: integer
          example: 5001
        sender:
          type: string
          example: "Maria"
        content:
          type: string
          description: Il testo del messaggio o l'identificatore della foto
          example: "Ehi, tutto bene?"
        type:
          type: string
          enum: [text, image]
          example: "text"
        status:
          type: string
          enum: [sent, received, read]
          description: sent=inviato (1 spunta grigia), received=ricevuto (2 grigie), read=letto (2 blu)
          example: "read"
        reactions:
          type: array
          description: Lista delle reazioni a questo messaggio
          items:
            type: object
            properties:
              emoji:
                type: string
                example: "👍"
              user:
                type: string
                example: "Luigi"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-04T15:35:00Z"